Fix type of total_steps in Tracking1D task
Ensure total_steps is explicitly cast to int dtype after ceiling operation to prevent potential type issues in downstream code.

+++ src/canns/task/tracking1d.py
@@ -49,7 +49,7 @@ class Tracking1D(BaseTask):
         self.current_step = 0
         self.time_step = time_step
         self.total_duration = sum(duration)
-        self.total_steps = u.math.ceil(self.total_duration / self.time_step)
+        self.total_steps = u.math.ceil(self.total_duration / self.time_step).astype(dtype=int)
 
         # Pre-computes the entire sequence of external inputs for the simulation.
         self.Iext_sequence = self._make_Iext_sequence()
